requirements.txt
# Core engine
anthropic>=0.18.0
pyyaml>=6.0
rich>=13.0
snowflake-connector-python>=3.0.0
playwright>=1.40.0
pytest>=7.0.0

# LLM providers
openai>=1.0.0
google-generativeai>=0.3.0

# FastAPI backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
cryptography>=41.0.0
resend>=0.7.0
python-dotenv>=1.0.0
duckdb>=0.9.0
